Move POS preview helpers to app layer
