+++ WooCommerce/Classes/POS/Presentation/ItemCardView.swift
@@ -1,4 +1,3 @@
-import class Yosemite.POSProductProvider
 import protocol Yosemite.POSItem
 import SwiftUI
 
@@ -33,8 +32,7 @@ struct ItemCardView: View {
 
 #if DEBUG
 #Preview {
-    // TODO: https://github.com/woocommerce/woocommerce-ios/issues/12917
-    // Some Yosemite imports are only needed for previews
-    ItemCardView(item: POSProductProvider.provideProductForPreview(), onItemCardTapped: { })
+    ItemCardView(item: POSItemProviderPreview().providePointOfSaleItem(),
+                 onItemCardTapped: { })
 }
 #endif
