+++ WooCommerce/Classes/POS/Presentation/ItemGridView.swift
@@ -1,4 +1,3 @@
-import class Yosemite.POSProductProvider
 import SwiftUI
 
 struct ItemGridView: View {
@@ -49,9 +48,7 @@ private extension ItemGridView {
 
 #if DEBUG
 #Preview {
-    // TODO: https://github.com/woocommerce/woocommerce-ios/issues/12917
-    // The Yosemite imports are only needed for previews
-    ItemGridView(viewModel: PointOfSaleDashboardViewModel(items: POSProductProvider.provideProductsForPreview(),
-                                                             cardPresentPaymentService: CardPresentPaymentService(siteID: 0)))
+    ItemGridView(viewModel: PointOfSaleDashboardViewModel(items: POSItemProviderPreview().providePointOfSaleItems(),
+                                                          cardPresentPaymentService: CardPresentPaymentService(siteID: 0)))
 }
 #endif
