+++ WooCommerce/Classes/POS/Presentation/PointOfSaleDashboardView.swift
@@ -1,4 +1,3 @@
-import class Yosemite.POSProductProvider
 import SwiftUI
 
 struct PointOfSaleDashboardView: View {
@@ -92,9 +91,8 @@ fileprivate extension CardPresentPaymentEvent {
 
 #if DEBUG
 #Preview {
-    // TODO: https://github.com/woocommerce/woocommerce-ios/issues/12917
-    // The Yosemite imports are only needed for previews
-    PointOfSaleDashboardView(viewModel: PointOfSaleDashboardViewModel(items: POSProductProvider.provideProductsForPreview(),
-                                                                      cardPresentPaymentService: CardPresentPaymentService(siteID: 0)))
+    PointOfSaleDashboardView(
+        viewModel: PointOfSaleDashboardViewModel(items: POSItemProviderPreview().providePointOfSaleItems(),
+                                                 cardPresentPaymentService: CardPresentPaymentService(siteID: 0)))
 }
 #endif
